+++ src/constants/debate_template.ts
